feature(client; server): reworked clients and servers classes feature(maui): added total bytes transferred info to notification feature(maui): added exit button to notification doc(readme): updated

+++ src/MauiClient/Pages/TunnelEditPage.xaml.cs
@@ -24,7 +24,17 @@ public partial class TunnelEditPage : CContentPage
 
   private async void DeleteBtn_Clicked(object _sender, EventArgs _e)
   {
+    var confirmed = await DisplayAlert(
+      "Confirmation",
+      $"Are you sure you want to delete '{p_tunnelModel.Name}' tunnel?",
+      "Yes",
+      "Cancel");
+
+    if (!confirmed)
+      return;
+
     p_tunnelsConfCtrl.DeleteTunnelConf(p_tunnelModel.Guid);
     _ = await Navigation.PopAsync();
   }
+
 }
